refactor: moved constants & chains to proper files
