+++ packages/keplr/src/index.ts
@@ -1,6 +1,8 @@
 import { Keplr } from "@keplr-wallet/types";
 
 export * from "./signer";
+export * from "./constants";
+export * from "./chains";
 
 // Export the Keplr types
 export { Keplr };
