+++ packages/keplr/src/signer.ts
@@ -9,6 +9,7 @@ import { SignDoc } from "cosmjs-types/cosmos/tx/v1beta1/tx";
 import { AminoSignResponse, StdSignDoc, Algo } from "@cosmjs/amino";
 import { assert } from "@cosmjs/utils";
 import { Signer, SignerStatus, SigningMode } from "@desmoslabs/desmjs";
+import { DesmosMainnet } from "./chains";
 
 declare global {
   interface Window extends KeplrWindow {}
@@ -19,95 +20,6 @@ export interface KeplrSignerOptions {
   chainInfo: ChainInfo;
   signOptions: KeplrSignOptions;
 }
-
-export const DesmosMainnet: ChainInfo = {
-  chainId: "desmos-mainnet",
-  chainName: "Desmos",
-  rpc: "https://rpc.mainnet.desmos.network",
-  rest: "https://api.mainnet.desmos.network",
-  bip44: {
-    coinType: 852,
-  },
-  bech32Config: {
-    bech32PrefixAccAddr: "desmos",
-    bech32PrefixAccPub: "desmospub",
-    bech32PrefixValAddr: "desmosvaloper",
-    bech32PrefixValPub: "desmosvaloperpub",
-    bech32PrefixConsAddr: "desmosvalcons",
-    bech32PrefixConsPub: "desmosvalconspub",
-  },
-  currencies: [
-    {
-      coinDenom: "DSM",
-      coinMinimalDenom: "udsm",
-      coinDecimals: 6,
-      coinGeckoId: "desmos",
-    },
-  ],
-  feeCurrencies: [
-    {
-      coinDenom: "udsm",
-      coinMinimalDenom: "udsm",
-      coinDecimals: 6,
-      coinGeckoId: "desmos",
-    },
-  ],
-  stakeCurrency: {
-    coinDenom: "DSM",
-    coinMinimalDenom: "udsm",
-    coinDecimals: 6,
-    coinGeckoId: "desmos",
-  },
-  gasPriceStep: {
-    low: 0.01,
-    average: 0.03,
-    high: 0.05,
-  },
-  features: ["stargate", "ibc-transfer", "no-legacy-stdTx", "ibc-go"],
-};
-
-export const DesmosTestnet: ChainInfo = {
-  chainId: "morpheus-apollo-2",
-  chainName: "Desmos Testnet",
-  rpc: "https://rpc.morpheus.desmos.network",
-  rest: "https://lcd.morpheus.desmos.network",
-  bip44: {
-    coinType: 852,
-  },
-  bech32Config: {
-    bech32PrefixAccAddr: "desmos",
-    bech32PrefixAccPub: "desmospub",
-    bech32PrefixValAddr: "desmosvaloper",
-    bech32PrefixValPub: "desmosvaloperpub",
-    bech32PrefixConsAddr: "desmosvalcons",
-    bech32PrefixConsPub: "desmosvalconspub",
-  },
-  currencies: [
-    {
-      coinDenom: "DARIC",
-      coinMinimalDenom: "udaric",
-      coinDecimals: 6,
-    },
-  ],
-  feeCurrencies: [
-    {
-      coinDenom: "udaric",
-      coinMinimalDenom: "udaric",
-      coinDecimals: 6,
-    },
-  ],
-  stakeCurrency: {
-    coinDenom: "DARIC",
-    coinMinimalDenom: "udaric",
-    coinDecimals: 6,
-  },
-  gasPriceStep: {
-    low: 0.01,
-    average: 0.03,
-    high: 0.05,
-  },
-  features: ["stargate", "ibc-transfer", "no-legacy-stdTx", "ibc-go"],
-};
 
 /**
  * Signer that use Keplr to sign a transaction.
